// workspace/projects/porting/src-ide/app/service/report/Interface.ts
/*
 * Copyright 2022 Huawei Technologies Co., Ltd
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

interface Item {
    value: Array<string> | string;
    label: string;
    isSuccessed?: string;
    fixSum?: number; // 可修改总数
}

interface RightItem {
    title: string;
    value: number;
}

interface ScanItems {
    soFile: {
        label: string
    };
    type: Array<any>;
}

interface TextForm {
    firstItem: Item;
    fifthItem?: Item;
    sixthItem?: Item;
    seventhItem?: Item;
}


export {
    RightItem,
    ScanItems,
    TextForm
};